+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Code freeze date: YYYY-MM-DD
 
 ### Changed
 
+- Module `climada_petals.hazard.river_flood`, class `RiverFlood`: New method `from_aqueduct_tif` to read river flood data directly from the online Aqueduct tif files, and renamed methods `from_nc` to `from_isimip_nc` and `set_from_nc` to `set_from_isimip_nc` [#108](https://github.com/CLIMADA-project/climada_petals/pull/108)
+
 ### Fixed
 
 ### Deprecated